lure/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lure"
version = "1.0.0"
description = "Phishing Analysis & IOC Extraction Platform for L1/L2 SOC Analysts"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # Email parsing
    "mail-parser>=3.15",
    "extract-msg>=0.48",
    # Auth validation
    "pyspf>=2.0",
    "dkimpy>=1.1",
    "checkdmarc>=5.0",
    "dnspython>=2.4",
    # IOC extraction
    "iocextract>=1.16",
    "iocsearcher>=0.5",
    "ioc-fanger>=1.2",
    "beautifulsoup4>=4.12",
    "pdfplumber>=0.10",
    "python-docx>=1.1",
    "oletools[full]>=0.60",
    "urlextract>=1.8",
    "tldextract>=5.1",
    # Config & CLI
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "python-dotenv>=1.0",
    "typer>=0.9",
    "rich>=13.7",
    # Async + HTTP
    "aiohttp>=3.9",
    "httpx>=0.26",
    "requests>=2.31",
    # File type detection
    "python-magic>=0.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.28",
    "responses>=0.24",
]
yara = ["yara-python>=4.5", "yara-mail>=1.3"]
archive = ["py7zr>=0.21", "rarfile>=4.1"]
enrichment = ["vt-py>=0.18", "shodan>=1.31", "OTXv2>=1.5", "ipwhois>=1.2", "python-whois>=0.9"]
output = ["stix2>=3.0", "pymisp>=2.4", "thehive4py>=1.8", "weasyprint>=61", "jinja2>=3.1"]
llm = ["instructor>=0.6", "ollama>=0.1"]
all = ["lure[dev,yara,archive,enrichment,output,llm]"]

[project.scripts]
lure = "lure.cli:app"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
